requirements.txt
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
langchain==0.1.0
langchain-google-genai==0.0.5
langchain-community==0.0.10
chromadb==0.4.22
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.3
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pandas==2.1.4
scikit-learn==1.3.2
lxml==4.9.3
pdfplumber==0.10.3
PyPDF2==3.0.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
